Fix compile warnings
Former-commit-id: 8402cc03d2cba5e1c2bb711971ce7a6770587318

+++ src/audio.cpp
@@ -274,7 +274,7 @@ SoundHandle Audio::playSound3D(Sound* sound, SoundType soundType,
 
     if (loop)
     {
-        ps.playPosition = random(randomSeries, 0.f, sound->numSamples);
+        ps.playPosition = random(randomSeries, 0.f, (f32)sound->numSamples);
     }
 
     std::lock_guard<std::mutex> lock(audioMutex);
